pyproject.toml
[project]
name = "status-monitor"
version = "0.1.0"
description = "Event-driven status page monitor using webhooks"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
